+++ packages/bridge-controller/src/utils/fetch.ts
@@ -2,11 +2,16 @@ import { StructError } from '@metamask/superstruct';
 import type { CaipAssetType, CaipChainId, Hex } from '@metamask/utils';
 import { Duration } from '@metamask/utils';
 
+import { isBitcoinChainId } from './bridge';
 import {
   formatAddressToCaipReference,
   formatChainIdToDec,
 } from './caip-formatters';
-import { validateQuoteResponse, validateSwapsTokenObject } from './validators';
+import {
+  validateQuoteResponse,
+  validateBitcoinQuoteResponse,
+  validateSwapsTokenObject,
+} from './validators';
 import type {
   QuoteResponse,
   FetchFunction,
@@ -127,6 +132,11 @@ export async function fetchBridgeQuotes(
   const filteredQuotes = quotes.filter(
     (quoteResponse: unknown): quoteResponse is QuoteResponse => {
       try {
+        const isBitcoinQuote = isBitcoinChainId(request.srcChainId);
+
+        if (isBitcoinQuote) {
+          return validateBitcoinQuoteResponse(quoteResponse);
+        }
         return validateQuoteResponse(quoteResponse);
       } catch (error) {
         if (error instanceof StructError) {